Adding documentation and styling.
Functionality to create a custom menu and trigger a cloud function
through pubsub messages based on the menu item selected.

Adding documentation, styling and splitting files in logical modules.

Applying the type declaration in JSDocs, because the scripts are going
to be deployed in the AppScript interpreter of Google Sheets. This only
accepts these Type declarations.

Change-Id: I89f69f103daae5c628cf1407b6e0f08cee52a319