Refactoring process_operations_and_errors.
In order to accept API response from Campaign Operations as wel as
Sitelink Operations. I modified the process_operations_and_errors
function in the utils file. It takes the sheetname as an extra argument,
which is used to write results back to the correct sheet and columns
respectively.

The tests and references have been updates to reflect this change.

Change-Id: I739b7746587f8ae8e06f2c694f3e279b44f5a1d2